requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
youtube-transcript-api
sentence-transformers
faiss-cpu
python-dotenv
